LPS-171250 Extract Then Card to its own component
