+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/ActionContainer.tsx
@@ -13,41 +13,22 @@
  */
 
 import {ActionError} from '..';
-import {ClayCheckbox, ClaySelect} from '@clayui/form';
-import ClayIcon from '@clayui/icon';
-import ClayLabel from '@clayui/label';
-import {ClayTooltipProvider} from '@clayui/tooltip';
 import {
 	API,
 	Card,
 	CodeEditor,
 	CustomItem,
 	Input,
-	SelectWithOption,
 	SidebarCategory,
-	SingleSelect,
 } from '@liferay/object-js-components-web';
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 
 import PredefinedValuesTable from '../PredefinedValuesTable';
 
 import './ActionBuilder.scss';
-import {
-	fetchObjectDefinitionFields,
-	fetchObjectDefinitions,
-} from '../fetchUtil';
+import {fetchObjectDefinitionFields} from '../fetchUtil';
 import {WarningStates} from './ActionBuilder';
-
-export type ObjectsOptionsList = Array<
-	(
-		| React.ComponentProps<typeof ClaySelect.Option>
-		| React.ComponentProps<typeof ClaySelect.OptGroup>
-	) & {
-		options?: Array<React.ComponentProps<typeof ClaySelect.Option>>;
-		type?: 'group';
-	}
->;
-
+import {ThenContainer} from './ThenContainer';
 interface ActionContainerProps {
 	currentObjectDefinitionFields: ObjectField[];
 	errors: ActionError;
@@ -83,31 +64,9 @@ export function ActionContainer({
 	validateExpressionURL,
 	values,
 }: ActionContainerProps) {
-	const [notificationTemplates, setNotificationTemplates] = useState<
-		CustomItem<number>[]
-	>([]);
-
 	const [relationships, setRelationships] = useState<
 		ObjectDefinitionsRelationship[]
 	>([]);
-
-	const [objectsOptions, setObjectOptions] = useState<ObjectsOptionsList>([]);
-
-	const notificationTemplateLabel = useMemo(() => {
-		return notificationTemplates.find(
-			({value}) => value === values.parameters?.notificationTemplateId
-		)?.label;
-	}, [notificationTemplates, values.parameters]);
-
-	const actionExecutors = useMemo(() => {
-		const executors = new Map<string, string>();
-
-		newObjectActionExecutors.forEach(({label, value}) => {
-			value && executors.set(value, label);
-		});
-
-		return executors;
-	}, [newObjectActionExecutors]);
 
 	const isValidField = ({
 		businessType,
@@ -200,49 +159,7 @@ export function ActionContainer({
 	};
 
 	useEffect(() => {
-		if (values.objectActionExecutorKey === 'notification') {
-			const makeFetch = async () => {
-				const NotificationTemplatesResponse = await API.getNotificationTemplates();
-
-				let notificationArray: NotificationTemplate[] = NotificationTemplatesResponse;
-
-				if (systemObject) {
-					notificationArray = NotificationTemplatesResponse.filter(
-						(notificationTemplate) =>
-							notificationTemplate.type !== 'userNotification'
-					);
-				}
-
-				setNotificationTemplates(
-					notificationArray.map(({id, name, type}) => ({
-						label: name,
-						type,
-						value: id,
-					}))
-				);
-			};
-
-			makeFetch();
-		}
-
-		if (values.objectActionExecutorKey === 'add-object-entry') {
-			fetchObjectDefinitions(
-				objectDefinitionsRelationshipsURL,
-				values,
-				setRelationships,
-				setObjectOptions
-			);
-
-			fetchObjectDefinitionFields(
-				objectDefinitionId,
-				objectDefinitionExternalReferenceCode,
-				values,
-				isValidField,
-				setCurrentObjectDefinitionFields,
-				setValues
-			);
-		}
-		else if (values.objectActionExecutorKey === 'update-object-entry') {
+		if (values.objectActionExecutorKey === 'update-object-entry') {
 			updateParameters(objectDefinitionExternalReferenceCode);
 			fetchObjectDefinitionFields(
 				objectDefinitionId,
@@ -255,146 +172,36 @@ export function ActionContainer({
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [
-		values.objectActionExecutorKey,
-		systemObject,
 		objectDefinitionId,
 		objectDefinitionExternalReferenceCode,
 		objectDefinitionsRelationshipsURL,
+		systemObject,
+		values.objectActionExecutorKey,
 	]);
 
 	return (
 		<Card title={Liferay.Language.get('action')}>
-			<Card
-				title={Liferay.Language.get('then[object]')}
-				viewMode="inline"
-			>
-				<div className="lfr-object__action-builder-then">
-					<SingleSelect
-						error={errors.objectActionExecutorKey}
-						onChange={({value}) => {
-							if (value === 'add-object-entry') {
-								fetchObjectDefinitions(
-									objectDefinitionsRelationshipsURL,
-									values,
-									setRelationships,
-									setObjectOptions
-								);
-							}
-							setValues({
-								objectActionExecutorKey: value,
-								parameters: {},
-							});
-						}}
-						options={
-							Liferay.FeatureFlags['LPS-153714']
-								? newObjectActionExecutors
-								: objectActionExecutors
-						}
-						placeholder={Liferay.Language.get('choose-an-action')}
-						value={actionExecutors.get(
-							values.objectActionExecutorKey ?? ''
-						)}
-					/>
-
-					{values.objectActionExecutorKey === 'add-object-entry' && (
-						<>
-							on
-							<SelectWithOption
-								aria-label={Liferay.Language.get(
-									'choose-an-object'
-								)}
-								error={
-									errors.objectDefinitionExternalReferenceCode
-								}
-								onChange={({target: {value}}) =>
-									updateParameters(value)
-								}
-								options={objectsOptions}
-								value={
-									values.parameters
-										?.objectDefinitionExternalReferenceCode
-										? `${values.parameters.objectDefinitionExternalReferenceCode},${values.parameters.objectDefinitionId}`
-										: ''
-								}
-							/>
-							{values.parameters?.relatedObjectEntries !==
-								undefined && (
-								<>
-									<ClayCheckbox
-										checked={
-											values.parameters
-												.relatedObjectEntries === true
-										}
-										disabled={false}
-										label={Liferay.Language.get(
-											'also-relate-entries'
-										)}
-										onChange={({target: {checked}}) => {
-											setValues({
-												parameters: {
-													...values.parameters,
-													relatedObjectEntries: checked,
-												},
-											});
-										}}
-									/>
-									<ClayTooltipProvider>
-										<div
-											data-tooltip-align="top"
-											title={Liferay.Language.get(
-												'automatically-relate-object-entries-involved-in-the-action'
-											)}
-										>
-											<ClayIcon
-												className=".lfr-object__action-builder-tooltip-icon"
-												symbol="question-circle-full"
-											/>
-										</div>
-									</ClayTooltipProvider>
-								</>
-							)}
-						</>
-					)}
-
-					{values.objectActionExecutorKey === 'notification' && (
-						<SingleSelect<CustomItem<number>>
-							className="lfr-object__action-builder-notification-then"
-							error={errors.objectActionExecutorKey}
-							onChange={({value}) => {
-								setValues({
-									parameters: {
-										...values.parameters,
-										notificationTemplateId: value,
-									},
-								});
-							}}
-							options={notificationTemplates}
-							required
-							value={notificationTemplateLabel}
-						>
-							{notificationTemplates.map(
-								(option) =>
-									Liferay.FeatureFlags['LPS-162133'] &&
-									option.type && (
-										<ClayLabel
-											displayType={
-												option.type === 'email'
-													? 'success'
-													: 'info'
-											}
-										>
-											{option.type === 'email'
-												? Liferay.Language.get('email')
-												: Liferay.Language.get(
-														'user-notification'
-												  )}
-										</ClayLabel>
-									)
-							)}
-						</SingleSelect>
-					)}
-				</div>
-			</Card>
+			<ThenContainer
+				errors={errors}
+				isValidField={isValidField}
+				newObjectActionExecutors={newObjectActionExecutors}
+				objectActionExecutors={objectActionExecutors}
+				objectDefinitionExternalReferenceCode={
+					objectDefinitionExternalReferenceCode
+				}
+				objectDefinitionId={objectDefinitionId}
+				objectDefinitionsRelationshipsURL={
+					objectDefinitionsRelationshipsURL
+				}
+				setCurrentObjectDefinitionFields={
+					setCurrentObjectDefinitionFields
+				}
+				setRelationships={setRelationships}
+				setValues={setValues}
+				systemObject={systemObject}
+				updateParameters={updateParameters}
+				values={values}
+			/>
 
 			{(values.objectActionExecutorKey === 'add-object-entry' ||
 				values.objectActionExecutorKey === 'update-object-entry') &&
